@@ -1,4 +1,4 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
 """
 vg_msga.py: use vg's banded aligner to align contigs into the graph.  By associating BED regions with each
 contig, we can parallelize by chromosome. 
@@ -7,9 +7,8 @@ contig, we can parallelize by chromosome.
 from __future__ import print_function
 import argparse, sys, os, os.path, errno, random, subprocess, shutil, itertools, glob, tarfile
 import doctest, re, json, collections, time, timeit
-import logging, logging.handlers, SocketServer, struct, socket, threading
+import logging, logging.handlers, struct, socket, threading
 import string
-import urlparse
 import getpass
 import pdb
 import gzip
