@@ -1,4 +1,4 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
 """
 vg_plot.py: Make plots from toil-vg experiments
 
@@ -6,9 +6,8 @@ vg_plot.py: Make plots from toil-vg experiments
 from __future__ import print_function
 import argparse, sys, os, os.path, errno, random, subprocess, shutil, itertools, glob, tarfile
 import doctest, re, json, collections, time, timeit
-import logging, logging.handlers, SocketServer, struct, socket, threading
+import logging, logging.handlers, struct, socket, threading
 import string, math
-import urlparse
 import getpass
 import pdb
 import gzip
