@@ -1,4 +1,4 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
 """
 vg_toil.py: Run the mapping and variant calling on all the servers in
 parallel using the vg framework.
@@ -13,9 +13,8 @@ chr_length_list obtained from Mike Lin's vg dnanexus pipeline configuration
 from __future__ import print_function
 import argparse, sys, os, os.path, errno, random, subprocess, shutil, itertools, glob, tarfile
 import doctest, re, json, collections, time, timeit
-import logging, logging.handlers, SocketServer, struct, socket, threading
+import logging, logging.handlers, struct, socket, threading
 import string
-import urlparse
 import getpass
 import logging
 import pkg_resources
